Move interfaces into files of their own
