+++ libraries/String.class.php
@@ -1,136 +1,10 @@
 <?php
 /**
- * Defines a set of specialized string functions.
+ * Specialized string class for phpMyAdmin.
+ * The SQL Parser code relies heavily on these functions.
  *
  * @package PhpMyAdmin-String
- * @todo    May be move this into file of its own
  */
-interface PMA_StringType
-{
-    /**
-     * Checks if a character is an alphanumeric one
-     *
-     * @param string $c character to check for
-     *
-     * @return boolean whether the character is an alphanumeric one or not
-     */
-    public function isAlnum($c);
-
-    /**
-     * Checks if a character is an alphabetic one
-     *
-     * @param string $c character to check for
-     *
-     * @return boolean whether the character is an alphabetic one or not
-     */
-    public function isAlpha($c);
-
-    /**
-     * Checks if a character is a digit
-     *
-     * @param string $c character to check for
-     *
-     * @return boolean whether the character is a digit or not
-     */
-    public function isDigit($c);
-
-    /**
-     * Checks if a character is an upper alphabetic one
-     *
-     * @param string $c character to check for
-     *
-     * @return boolean whether the character is an upper alphabetic one or not
-     */
-    public function isUpper($c);
-
-
-    /**
-     * Checks if a character is a lower alphabetic one
-     *
-     * @param string $c character to check for
-     *
-     * @return boolean whether the character is a lower alphabetic one or not
-     */
-    public function isLower($c);
-
-    /**
-     * Checks if a character is a space one
-     *
-     * @param string $c character to check for
-     *
-     * @return boolean whether the character is a space one or not
-     */
-    public function isSpace($c);
-
-    /**
-     * Checks if a character is an hexadecimal digit
-     *
-     * @param string $c character to check for
-     *
-     * @return boolean whether the character is an hexadecimal digit or not
-     */
-    public function isHexDigit($c);
-
-    /**
-     * Checks if a number is in a range
-     *
-     * @param integer $num   number to check for
-     * @param integer $lower lower bound
-     * @param integer $upper upper bound
-     *
-     * @return boolean  whether the number is in the range or not
-     */
-    public function numberInRangeInclusive($num, $lower, $upper);
-}
-
-/**
- * Defines a set of specialized string functions.
- *
- * @package PhpMyAdmin-String
- * @todo    May be move this into file of its own
- */
-interface PMA_StringByte
-{
-    /**
-     * Returns length of string depending on current charset.
-     *
-     * @param string $string string to count
-     *
-     * @return int string length
-     */
-
-    public function strlen($string);
-
-    /**
-     * Returns substring from string, works depending on current charset.
-     *
-     * @param string $string string to count
-     * @param int    $start  start of substring
-     * @param int    $length length of substring
-     *
-     * @return string the sub string
-     */
-    public function substr($string, $start, $length = 2147483647);
-
-    /**
-     * Returns postion of $needle in $haystack or false if not found
-     *
-     * @param string $haystack the string being checked
-     * @param string $needle   the string to find in haystack
-     * @param int    $offset   the search offset
-     *
-     * @return integer position of $needle in $haystack or false
-     */
-    public function strpos($haystack, $needle, $offset = 0);
-    /**
-     * Make a string lowercase
-     *
-     * @param string $string the string being lowercased
-     *
-     * @return string the lower case string
-     */
-    public function strtolower($string);
-}
 
 /**
  * Specialized string class for phpMyAdmin.
