+++ libraries/StringAbstractType.class.php
@@ -9,7 +9,7 @@ if (! defined('PHPMYADMIN')) {
     exit;
 }
 
-require_once 'libraries/String.class.php';
+require_once 'libraries/StringType.int.php';
 
 /**
  * Implements PMA_StringByte interface using native PHP functions.
